First pass of refactoring buffer handling
