+++ jgltf-model-builder/src/main/java/de/javagl/jgltf/model/creation/AccessorModels.java
@@ -391,83 +391,6 @@ public class AccessorModels
         return accessorModel;
     }
     
-    /**
-     * Compute the number of bytes that the given {@link AccessorModel} data
-     * has to be aligned to. 
-     * 
-     * @param accessorModel The {@link AccessorModel}
-     * @return The alignment bytes
-     */
-    static int computeAlignmentBytes(AccessorModel accessorModel)
-    {
-        return accessorModel.getComponentSizeInBytes();
-    }
-    
-    /**
-     * Compute the alignment for the given {@link AccessorModel} instances.
-     * This is the least common multiple of the alignment of all instances.
-     * 
-     * @param accessorModels The {@link AccessorModel} instances
-     * @return The alignment
-     */
-    static int computeAlignmentBytes(
-        Iterable<? extends AccessorModel> accessorModels)
-    {
-        int alignmentBytes = 1;
-        for (AccessorModel accessorModel : accessorModels)
-        {
-            alignmentBytes = Utils.computeLeastCommonMultiple(alignmentBytes, 
-                AccessorModels.computeAlignmentBytes(accessorModel));
-        }
-        return alignmentBytes;
-    }
-    
-    /**
-     * Compute the byte stride that is common for the given 
-     * {@link AccessorModel} instances.
-     * 
-     * This is the maximum of the {@link AccessorModel#getElementSizeInBytes() 
-     * element sizes} and the {@link AccessorModel#getByteStride() byte 
-     * strides} of the given models. 
-     * 
-     * @param accessorModels The {@link AccessorModel} instances
-     * @return The common byte stride
-     */
-    private static int computeCommonByteStride(
-        Iterable<? extends AccessorModel> accessorModels)
-    {
-        int commonByteStride = 1;
-        for (AccessorModel accessorModel : accessorModels)
-        {
-            int elementSize = accessorModel.getPaddedElementSizeInBytes();
-            commonByteStride = Math.max(commonByteStride, elementSize);
-            
-            int byteStride = accessorModel.getByteStride();
-            commonByteStride = Math.max(commonByteStride, byteStride);
-        }
-        return commonByteStride;
-    }
-
-    /**
-     * Compute the byte stride that is common for the given 
-     * {@link AccessorModel} instances when they are used as
-     * vertex attributes. 
-     * 
-     * The byte stride for vertex attributes must be a multiple of 4. So the 
-     * result of this function is the smallest multiple of 4 that is at least 
-     * as large as the {@link #computeCommonByteStride(Iterable) common byte 
-     * stride}. 
-     * 
-     * @param accessorModels The {@link AccessorModel} instances
-     * @return The common byte stride
-     */
-    static int computeCommonVertexAttributeByteStride(
-        Iterable<? extends AccessorModel> accessorModels)
-    {
-        int commonByteStride = computeCommonByteStride(accessorModels);
-        return Utils.pad(commonByteStride, 4);
-    }
-    
     
     /**
      * Private constructor to prevent instantiation
