+++ jgltf-model/src/main/java/de/javagl/jgltf/model/io/v2/BinaryAssetCreatorV2.java
@@ -29,20 +29,16 @@ package de.javagl.jgltf.model.io.v2;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
 
 import de.javagl.jgltf.impl.v2.Buffer;
 import de.javagl.jgltf.impl.v2.BufferView;
 import de.javagl.jgltf.impl.v2.GlTF;
 import de.javagl.jgltf.impl.v2.Image;
-import de.javagl.jgltf.model.BufferModel;
+import de.javagl.jgltf.model.BufferViewModel;
 import de.javagl.jgltf.model.GltfModel;
 import de.javagl.jgltf.model.ImageModel;
-import de.javagl.jgltf.model.Optionals;
 import de.javagl.jgltf.model.io.Buffers;
 import de.javagl.jgltf.model.io.MimeTypes;
 import de.javagl.jgltf.model.v2.GltfCreatorV2;
@@ -71,8 +67,7 @@ final class BinaryAssetCreatorV2
     /**
      * Create a binary {@link GltfAssetV2} from the given {@link GltfModel}.
      * The resulting asset will have a {@link GlTF} that uses references to
-     * {@link BufferView} objects in its {@link Buffer} and {@link Image}
-     * elements. 
+     * {@link BufferView} objects in its {@link Image} elements. 
      * 
      * @param gltfModel The {@link GltfModel}
      * @return The {@link GltfAssetV2}
@@ -81,80 +76,72 @@ final class BinaryAssetCreatorV2
     {
         GlTF outputGltf = GltfCreatorV2.create(gltfModel);
         
-        // Create the new byte buffer for the data of the "binary_glTF" Buffer
-        int binaryGltfBufferSize = 
-            computeBinaryGltfBufferSize(gltfModel);
-        ByteBuffer binaryGltfByteBuffer = 
-            Buffers.create(binaryGltfBufferSize);
-
-        // Create the binary Buffer, 
-        Buffer binaryGltfBuffer = new Buffer();
-        binaryGltfBuffer.setByteLength(binaryGltfBufferSize);
-        outputGltf.setBuffers(Collections.singletonList(binaryGltfBuffer));
-
-        // Create a defensive copy of the original image list
-        List<Image> oldImages = copy(outputGltf.getImages());
-
-        // Place the data from buffers and images into the new binary glTF 
-        // buffer. The mappings from IDs to offsets inside the resulting 
-        // buffer will be used to compute the offsets for the buffer views
-        List<ByteBuffer> bufferDatas = 
-            gltfModel.getBufferModels().stream()
-            .map(BufferModel::getBufferData)
-            .collect(Collectors.toList());
-        Map<Integer, Integer> bufferOffsets = concatBuffers(
-            bufferDatas, binaryGltfByteBuffer);
-        List<ByteBuffer> imageDatas = 
-            gltfModel.getImageModels().stream()
-            .map(ImageModel::getImageData)
-            .collect(Collectors.toList());
-        Map<Integer, Integer> imageOffsets = concatBuffers(
-            imageDatas, binaryGltfByteBuffer);
-        binaryGltfByteBuffer.position(0);
-
-        // For all existing BufferViews, create new ones that are updated to 
-        // refer to the new binary glTF buffer, with the appropriate offset
+        // The data from all buffer views and all images has to be put
+        // into a single Buffer (namely, the buffer whose ByteBuffer
+        // will become the GltfAssetV2#getBinaryData)
+        List<ByteBuffer> bufferViewDatas = new ArrayList<ByteBuffer>();
+        int byteOffset = 0;
+        
+        // For each old buffer view: Create a new buffer view that
+        // is updated to refer to the new binary glTF buffer,
+        // and collect its data
         List<BufferView> oldBufferViews = 
             copy(outputGltf.getBufferViews());
         List<BufferView> newBufferViews = 
             new ArrayList<BufferView>();
+        List<BufferViewModel> bufferViewModels = 
+            gltfModel.getBufferViewModels();
         for (int i = 0; i < oldBufferViews.size(); i++)
         {
             BufferView oldBufferView = oldBufferViews.get(i);
             BufferView newBufferView = GltfUtilsV2.copy(oldBufferView);
 
             newBufferView.setBuffer(0);
-            Integer oldBufferIndex = oldBufferView.getBuffer();
-            int oldByteOffset = Optionals.of(oldBufferView.getByteOffset(), 0);
-            int bufferOffset = bufferOffsets.get(oldBufferIndex);
-            int newByteOffset = oldByteOffset + bufferOffset;
-            newBufferView.setByteOffset(newByteOffset);
-
+            newBufferView.setByteOffset(byteOffset);
             newBufferViews.add(newBufferView);
+            
+            BufferViewModel bufferViewModel = bufferViewModels.get(i);
+            ByteBuffer bufferViewData = bufferViewModel.getBufferViewData();
+            bufferViewDatas.add(bufferViewData);
+            byteOffset += bufferViewData.capacity();
         }
-
-        // For all existing Images, create new ones that are updated to 
-        // refer to the new binary glTF buffer, using a bufferView 
-        // index that refers to a newly created BufferView
+        
+        // For all existing images: If the image does not yet refer to a 
+        // buffer view, then create a new buffer view and update the
+        // image accordingly. If the image already referred to a buffer view,
+        // then its data has already been collected, and the image can
+        // be kept as-it-is.
+        List<Image> oldImages = copy(outputGltf.getImages());
         List<Image> newImages = new ArrayList<Image>();
+        List<ImageModel> imageModels = gltfModel.getImageModels();
         for (int i = 0; i < oldImages.size(); i++)
         {
             Image oldImage = oldImages.get(i);
             Image newImage = GltfUtilsV2.copy(oldImage);
-
-            // Create the BufferView for the image
-            ImageModel imageModel = gltfModel.getImageModels().get(i);
+            
+            ImageModel imageModel = imageModels.get(i);
             ByteBuffer imageData = imageModel.getImageData();
-            int byteLength = imageData.capacity();
-            int byteOffset = imageOffsets.get(i);
-            BufferView imageBufferView = new BufferView();
-            imageBufferView.setBuffer(0);
-            imageBufferView.setByteOffset(byteOffset);
-            imageBufferView.setByteLength(byteLength);
 
-            int newBufferViewIndex = newBufferViews.size();
-            newImage.setBufferView(newBufferViewIndex);
-            newImage.setUri(null);
+            // If the image does not already refer to a buffer view,
+            // then create a new buffer view, update the image, and
+            // collect the data of the new buffer view
+            BufferViewModel imageBufferViewModel = 
+                imageModel.getBufferViewModel();
+            if (imageBufferViewModel == null)
+            {
+                bufferViewDatas.add(imageData);
+                byteOffset += imageData.capacity();
+                
+                BufferView imageBufferView = new BufferView();
+                imageBufferView.setBuffer(0);
+                imageBufferView.setByteOffset(byteOffset);
+                imageBufferView.setByteLength(imageData.capacity());
+                
+                int newBufferViewIndex = newBufferViews.size();
+                newImage.setBufferView(newBufferViewIndex);
+                newImage.setUri(null);
+                newBufferViews.add(imageBufferView);
+            }
             
             String imageMimeTypeString =
                 MimeTypes.guessImageMimeTypeString(
@@ -167,14 +154,19 @@ final class BinaryAssetCreatorV2
             {
                 newImage.setMimeType(imageMimeTypeString);
             }
-            
-            newBufferViews.add(imageBufferView);
             newImages.add(newImage);
         }
+        
+        // Create the binary Buffer, combining the data of all buffer views
+        // that have been collected or newly created for the images
+        Buffer binaryGltfBuffer = new Buffer();
+        ByteBuffer binaryGltfByteBuffer = Buffers.concat(bufferViewDatas);
+        binaryGltfBuffer.setByteLength(binaryGltfByteBuffer.capacity());
+        outputGltf.setBuffers(Collections.singletonList(binaryGltfBuffer));
 
         // Place the newly created lists into the output glTF,
         // if there have been non-null lists for them in the input
-        if (!oldImages.isEmpty())
+        if (!newImages.isEmpty())
         {
             outputGltf.setImages(newImages);
         }
@@ -184,55 +176,6 @@ final class BinaryAssetCreatorV2
         }
         return new GltfAssetV2(outputGltf, binaryGltfByteBuffer);
     }
-
-
-    /**
-     * Compute the total size that is required for the binary glTF buffer
-     * for the given {@link GltfModel}, which is the sum of all buffer
-     * sizes of all buffers and images
-     * 
-     * @param gltfModel The {@link GltfModel}
-     * @return The total size for the binary glTF buffer
-     */
-    private static int computeBinaryGltfBufferSize(GltfModel gltfModel)
-    {
-        int binaryGltfBufferSize = 0;
-        for (BufferModel bufferModel : gltfModel.getBufferModels())
-        {
-            ByteBuffer bufferData = bufferModel.getBufferData();
-            binaryGltfBufferSize += bufferData.capacity();
-        }
-        for (ImageModel imageModel : gltfModel.getImageModels())
-        {
-            ByteBuffer imageData = imageModel.getImageData();
-            binaryGltfBufferSize += imageData.capacity();
-        }
-        return binaryGltfBufferSize;
-    }
-
-    /**
-     * Put the contents of all byte buffers of the given list into the given 
-     * target buffer. This method assumes that the target buffer has a 
-     * sufficient capacity to hold all buffers.
-     * 
-     * @param buffers The buffers
-     * @param targetBuffer The target buffer
-     * @return A mapping from each key to the offset inside the target buffer
-     */
-    private static Map<Integer, Integer> concatBuffers(
-        List<? extends ByteBuffer> buffers, ByteBuffer targetBuffer)
-    {
-        Map<Integer, Integer> offsets = new LinkedHashMap<Integer, Integer>();
-        for (int i = 0; i < buffers.size(); i++)
-        {
-            ByteBuffer oldByteBuffer = buffers.get(i);
-            int offset = targetBuffer.position();
-            offsets.put(i, offset);
-            targetBuffer.put(oldByteBuffer.slice());
-        }
-        return offsets;
-    }
-
 
     /**
      * Creates a copy of the given list. If the given list is <code>null</code>,
